backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-chess==1.999
celery==5.3.6
redis==5.0.3
firebase-admin==6.5.0
requests==2.31.0
stripe==8.1.0
openai==1.14.2
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.3
